% $ Id: $
\name{grossRev}
\alias{grossRev}

\title{calculate gross revenue from 5 species size structured quota dynamic state variable model}
\description{

	The \code{grossRev} function returns gross revenue from a dynstate object 

}

\usage{	grossRev(dynstate)

}

\arguments{

  \item{dynstate}{A DynState object.}

}

\details{

	This function returns an object of class numeric with the calculated gross revenue from a dynstate model. Gross revenue is defined as the sum of landings * prices for all species and sizes.   

   }

\value{

    An object of class numeric.

}

\references{}

\author{Jan Jaap Poos}

\note{}

\seealso{}

\examples{}

\keyword{models} 
